* Adding scoping for managed temporaries in with statement
When with is done over a function call, it is copied into a temp
variable. Managed temps are usually created and finalized on entry or
exit of a function. This commit changes it so that the variable will be
initialized at the start of the with block and finalized once the with
block is left.

* nbas.pas: Adding new temporary option indicating it is manually
  managed
* ncgbas.pas: if marked as manually managed, don't generate init and
  final code
* pstatement.pas: add manual management for managed temporaries in with
  statement
* ppu.pas: Increse PPU version