* Adding ability to make Records non Copy or non Adref able
This allows the user to delete the copy and addref operator for a
record:
```pascal
type
  TMyRec = record
    ...
    class operator copy = nil;
    class operator addref = nil;
  end.
```
This makes the record non copyable or non addrefable. If the record is
non copyiable, any assignment which is not from a function result (and
therefore a temporary variable) is prevented by the compiler. If non
adrefable, then passing the type as a non const/var/out param is
prohibited.